requirements.txt
python-telegram-bot>=21.0
httpx>=0.27.0
APScheduler>=3.10.0
beautifulsoup4>=4.12.0
aiosqlite>=0.19.0
